@@ -1,23 +1,22 @@
 {
   "name": "cd-messenger",
-  "version": "2.7.16",
+  "version": "2.7.17-33",
   "description": "console log logger gulp notification browser node message",
   "main": "index.js",
   "reveal": true,
   "scripts": {
-    "all": "npm run eslint",
-    "build:dev": "node -r babel-register node_modules/.bin/webpack --hide-modules --config=webpack.config.babel.js && bash ./scripts/copy.sh",
-    "build:prod": "NODE_ENV=production node -r babel-register node_modules/.bin/webpack --hide-modules --config=webpack.config.babel.js && bash ./scripts/copy.sh",
-    "build:watch_temp": "node -r babel-register node_modules/.bin/webpack --hide-modules -w --config=webpack.config.babel.js",
-    "build:watch": "echo \"\n\n  ==> NOTE:  Running this causes an infinite loop, review cd-datetime-picker project for solution (gulp?)\n\n\"",
-    "build": "npm run clean && npm run build:dev && npm run build:prod",
-    "clean": "./node_modules/.bin/rimraf lib && ./node_modules/.bin/rimraf examples/lib",
+    "build": "npm run clean && npm run eslint && npm run build:dev && npm run build:prod",
+    "build:dev": "node -r babel-register node_modules/.bin/webpack --hide-modules --config=webpack.config.dev.js && bash ./scripts/copy.sh",
+    "build:prod": "NODE_ENV=production node -r babel-register node_modules/.bin/webpack --hide-modules --config=webpack.config.prod.js",
+    "build:watch": "node -r babel-register node_modules/.bin/webpack --hide-modules --config=webpack.config.dev.js --watch && bash ./scripts/copy.sh",
+    "clean": "./node_modules/.bin/rimraf dist && ./node_modules/.bin/rimraf examples/lib",
     "eslint": "eslint \"./**/*.js\"",
-    "lint": "npm run eslint",
-    "publish": "npm run build && bash ./scripts/delete-sourcemaps.sh",
+    "lint": "npm run eslint --silent",
+    "publish": "npm run build",
     "test:node": "node examples/node-test",
-    "test:watch": "./node_modules/.bin/mocha --compilers js:babel-core/register -w",
-    "test": "./node_modules/.bin/mocha --compilers js:babel-core/register"
+    "test:watch": "mocha --compilers js:babel-core/register -w",
+    "test": "mocha --compilers js:babel-core/register",
+    "test:coverage": "nyc --report-dir ./test/coverage -r html mocha --require babel-register && open ./test/coverage/index.html"
   },
   "repository": {
     "type": "git",
@@ -44,26 +43,28 @@
   "homepage": "https://github.com/mikeerickson/cd-messenger#readme",
   "devDependencies": {
     "@slightlytyler/webpack-shell-plugin": "0.4.5",
-    "babel-core": "6.22.1",
-    "babel-loader": "6.2.10",
-    "babel-preset-babili": "0.0.10",
+    "babel-core": "6.23.1",
+    "babel-loader": "6.4.0",
+    "babel-preset-babili": "0.0.12",
     "babel-preset-es2015-native-modules": "6.9.4",
     "babel-preset-es2015-without-strict": "0.0.4",
-    "babili-webpack-plugin": "0.0.9",
+    "babili-webpack-plugin": "0.0.11",
     "bump-version": "0.5.0",
     "chai": "3.5.0",
-    "eslint": "3.14.1",
-    "eslint-loader": "1.6.1",
+    "eslint": "3.17.1",
+    "eslint-loader": "1.6.3",
     "html-webpack-plugin": "2.28.0",
     "json-loader": "0.5.4",
     "lodash": "4.17.4",
     "mocha": "3.2.0",
     "mocha-sinon": "1.1.6",
     "progress-bar-webpack-plugin": "1.9.3",
-    "rimraf": "2.5.4",
+    "rimraf": "2.6.1",
     "sinon": "1.17.7",
-    "webpack": "2.2.0",
-    "webpack-build-notifier": "0.1.13"
+    "watch-ignore-webpack-plugin": "1.0.0",
+    "webpack": "2.2.1",
+    "webpack-build-notifier": "0.1.13",
+    "webpack-merge": "4.0.0"
   },
   "dependencies": {
     "chalk": "1.1.3",
